[REFACTOR] DictSearchPage
